feature: Enable interuption of speech but client app
Add an endpoint to interupt the sentance level streaming of audio to client. Because StreamingResponse is difficult to interrupt,
the agent_status is used in the stream with an new interrupt endpoint that changes this status.
ii) Remove problem character filter from text procession prior to sending to STT
iii) Upgrade /agent_to_say and /talk_to_agent to yeild the MultipartResponse expected by the front-end. Multipart response was introduced to deal with viseme (json) and audio being returned.