Feature: Make avatar config setting a dictionary
Breaking change: Previously avater settings were limited to just voice. HOwever, to extend this to support
a sequence of animations defined in the avatar settings it was changed to a diction. Avatar sttings
now always returns a dictionary andnot a string.

Chore: add new command to MakeFile to connect to Mongosh inside docker