Feature: Add redis caching for visemes, audio and blendshaepes.
To improve latency on common queries, redis in-memory caching is added. The key is the voice_id and the text input.
It became apparent that follow on questions proved problematic as they relate to earlier parts of the conversation
As a work around any input containing pronouns it, its them, they never hit the cache. Need a smarter solution